Added basic support for uploading files
Added basic support to write into s3 resources. Unfortunately the amazon aws API doesn't allow to open an OutputStream and
let the client stream the content. Hence we are creating a temporary file and using that to upload the content. Have to
investigate if there is another more elegant way to stream files without having to use another api or write the request on ourselves
 probably with an RestTemplate