Use original resource loader to delegate non s3 resources. Resolves spring-cloud/spring-cloud-aws#68
Prior to this commit the current resource load has not been used by the
SimpleStorageResourceLoader, creating problems in some circumstances
in web application. The resource loader now uses the delegate one in any
case. As a side consequence, the resource loader is created very late in the
bean life cycle, in order to be able to receive the resource loader of the
application context.