Fix Acknowledgement Graceful Shutdown
BatchingAcknowledgementProcessor had a racing condition where if the processor was stopped while there where still messages to ack in the main ack queue, it would stop polling and fail to acknowledge such messages even if there was time left for it.

This commit fixes this and adds relevant test coverage.