Added trailing slash to parameter store examples (#1086)
The trailing slash is required to avoid a leading dot in the property name. For example, if the parameter is named `/config/spring/message` and the `spring.config.import` property is `aws-parameterstore:/config/spring`, then the resulting property in the Spring environment will be `.message`. Instead, the property should be `message`, so the `spring.config.import` property should be `aws-parameterstore:/config/spring/`

Fixes #1058

Co-authored-by: Nathan Clement <nclement@qvalent.com>